+++ src/lib/components/modal/select-contract/ContractListCard.tsx
@@ -10,31 +10,37 @@ import {
   MenuButton,
   MenuList,
   Spacer,
+  chakra,
 } from "@chakra-ui/react";
 import dayjs from "dayjs";
 import { MdMoreHoriz, MdMode, MdDelete } from "react-icons/md";
 
 import { EditList, RemoveList } from "../list";
-import { getListIcon } from "lib/data";
+import { getListIcon, INSTANTIATED_LIST_NAME } from "lib/data";
 import type { ContractListInfo } from "lib/stores/contract";
+import { formatSlugName } from "lib/utils";
 
-interface ListCardProps {
-  item: ContractListInfo;
-  handleListSelect: (value: string) => void;
-  isReadOnly: boolean;
-  showLastUpdated: boolean;
-}
-export const ListCard = ({
-  item,
-  handleListSelect,
-  isReadOnly,
-  showLastUpdated,
-}: ListCardProps) => {
-  const iconProps = {
+const StyledIcon = chakra(Icon, {
+  baseStyle: {
     boxSize: "4",
     display: "flex",
     alignItems: "center",
-  };
+  },
+});
+
+interface ContractListCardProps {
+  item: ContractListInfo;
+  handleListSelect: (value: string) => void;
+  isReadOnly: boolean;
+}
+
+export const ContractListCard = ({
+  item,
+  handleListSelect,
+  isReadOnly,
+}: ContractListCardProps) => {
+  const showLastUpdated = item.slug !== formatSlugName(INSTANTIATED_LIST_NAME);
+
   return (
     <LinkBox cursor="pointer">
       <Flex
@@ -95,16 +101,14 @@ export const ListCard = ({
               <EditList
                 list={{ label: item.name, value: item.slug }}
                 menuItemProps={{
-                  icon: <Icon as={MdMode} style={iconProps} color="gray.600" />,
+                  icon: <StyledIcon as={MdMode} color="gray.600" />,
                   children: "Edit list name",
                 }}
               />
               <RemoveList
                 list={{ label: item.name, value: item.slug }}
                 menuItemProps={{
-                  icon: (
-                    <Icon as={MdDelete} style={iconProps} color="error.light" />
-                  ),
+                  icon: <StyledIcon as={MdDelete} color="error.light" />,
                   children: "Remove list",
                 }}
               />