+++ src/lib/hooks/index.ts
@@ -4,3 +4,4 @@ export * from "./useToast";
 export * from "./useEndpoint";
 export * from "./useUserKey";
 export * from "./useDummyWallet";
+export * from "./useAddressType";